refactor watchlist.js into app.js
