+++ packages/stockflux-watchlist/src/App.css
@@ -28,3 +28,107 @@
   --transition-duration-quickest: 0.01s;
   --shadow-depth: 2px;
 }
+
+* {
+  box-sizing: border-box;
+  color: var(--text-color);
+}
+
+body {
+  background-color: var(--window-background-color);
+  font-family: var(--font-roboto-sans-serif);
+  color: var(--text-color);
+  user-select: none;
+  margin: 0;
+  border: 1px solid var(--dark-background);
+}
+
+.watchlist-name {
+  margin-top: auto;
+  margin-bottom: auto;
+}
+
+.watchlist-name {
+  background-color: inherit;
+  padding: 5px;
+  padding-right: 0;
+  border: none;
+  font-weight: var(--font-weight-bolder);
+  font-size: var(--font-xxxlarge);
+  margin-left: calc(var(--padding-large) - 10px);
+  font-size: var(--font-xxxlarge);
+  width: 182px;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+
+.watchlist {
+  --cumulative-border-width: 3px;
+  height: calc(
+    100vh - var(--bottom-tab-height) + var(--cumulative-border-width)
+  );
+  display: flex;
+  flex-direction: column;
+  overflow: hidden;
+}
+
+.scroll-wrapper {
+  flex: 1 1 auto;
+  position: relative;
+  overflow-y: overlay;
+  overflow-x: hidden;
+  transition: 0.6s all;
+}
+
+.icon-clickable {
+  cursor: pointer;
+}
+
+.no-watchlist {
+  --padding-almost-large: 12px;
+  padding: var(--padding-almost-large) var(--padding-large)
+    var(--padding-almost-large) var(--padding-xlarge);
+  color: var(--text-grey);
+  font-size: var(--font-medium);
+  font-weight: var(--font-weight-normal);
+}
+
+.card-wrapper {
+  cursor: pointer;
+  border-width: 2px 0;
+  max-height: calc(var(--card-height) * 2);
+  border-bottom: 2px solid var(--clr-divider-light);
+}
+
+.card-wrapper * {
+  transition: var(--transition-duration-quick) all;
+}
+
+.card-wrapper * .icon-right {
+  pointer-events: all;
+}
+
+.card-wrapper.dragging {
+  transition-delay: var(--transition-duration-quickest);
+  transition-property: max-height, padding;
+  max-height: 0;
+}
+
+.card-wrapper.dragging * {
+  opacity: 0;
+}
+
+.card-wrapper.dragOverBottom,
+.card-wrapper.dragOver,
+.card-wrapper.dragging {
+  background-color: rgba(41,233,227,0.2);
+}
+
+.card-wrapper.dragOver {
+  padding-top: var(--card-height);
+}
+
+.card-wrapper.dragOverBottom {
+  padding-bottom: var(--card-height);
+}
